requirements.txt
aiogram==3.13.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
uvloop==0.20.0; platform_system == 'Linux'

